chore: use barrel export for utils
