+++ src/routes/[network]/(account)/ram/buy/tokens/+page.svelte
@@ -15,7 +15,7 @@
 	import AssetInput from '$lib/components/input/asset.svelte';
 
 	import { BuyRAMState } from '../state.svelte.js';
-	import { preventDefault } from '$lib/utils.js';
+	import { preventDefault } from '$lib/utils';
 
 	const context = getContext<UnicoveContext>('state');
 	const { data } = $props();
